\name{countries}
\alias{countries}
\docType{data}
\title{Birth and death rates, continent and population for selected countries}
\description{
Data from World Health Organization database http://apps.who.int/gho/data/view.main.CBDR2040.
Based on the example from Grammar of Graphics by Leland Wilkinson.
}
\author{
 Przemyslaw Biecek [based on WHO data]
}
\examples{
\dontrun{
	library(maps)
	data(countries)
	head(countries)
}
}
\keyword{datasets}
