\name{galton}
\alias{galton}
\alias{pearson}
\docType{data}
\title{Galton's and Pearson's height data for parents and children in centimeters}
\description{

This is a data set used by Galton in 1885 to study the relationship between a parent's height and their childrens.
The secon dataset was used by Pearson to investigate regression. 
Height is transformed from inches to centimeters.
}
\author{
 Przemyslaw Biecek
}
\examples{
head(galton)
head(pearson)
}
\keyword{datasets}
