\name{getMillwardBrown}
\alias{getMillwardBrown}
\title{MillwardBrown Pool Results}
\description{
Download pool results from MillwardBrown website.  
}
\usage{
getMillwardBrown()
}
\value{
A dataset in the molten format with pool date, party and percent of votes.
}
\author{
Maciej Beresewicz [data extraction] 
Przemyslaw Biecek [data melting] 
}
\examples{
\dontrun{
 getMillwardBrown()
}
}
