\name{imiona_warszawa}
\alias{imiona_warszawa}
\docType{data}
\title{Statistics of Names in Warsaw}
\description{
Names of kids [girls and boys] newborn in Warsaw between 2004 and 2014.

Data collected from http://gorny.edu.pl/imiona/index.php
}
\author{
 Bartlomiej & Agata Gorny [keeping statistics at http://gorny.edu.pl/imiona/index.php]
 Krzysiek Trajkowski [web scapping]
 Przemyslaw Biecek [R convertion]
}
\examples{
head(imiona_warszawa)
}
\keyword{datasets}
