\name{mandatySejmik2014}
\alias{mandatySejmik2014}
\docType{data}
\title{Local government elections in Poland 2014}
\description{
Number of mandates / seats for main four parties collected in local goverment elections in Poland in 2014.

PSL, PiS, PO, SLD, Inne stand for main paries.

ProcentWaznychGlosow is the percent of valid votes while long and lat stands for voivodships centers.

Source: http://wybory2014.pkw.gov.pl/pl/
}
\author{
 Przemyslaw Biecek base on data from http://wybory2014.pkw.gov.pl/pl/
}
\examples{
head(mandatySejmik2014)
}
\keyword{datasets}
