\name{maturaExam}
\alias{maturaExam}
\docType{data}
\title{Results from Matura Exams in Poland for Math and Language for Years 2010-2015}
\description{
This dataset is created based on data from ZPD package, see \code{https://github.com/zozlak/ZPD} and \code{http://zpd.ibe.edu.pl/doku.php?id=obazie}.
Each row shows results for one person that takes matura exams in a given year. 
}
\author{
 Przemyslaw Biecek [based on IBE / ZPD data]
}
\examples{
\dontrun{
	data(maturaExam)
	head(maturaExam)
}
}
\keyword{datasets}
