\name{skiJumps2013}
\alias{skiJumps2013}
\alias{skiJumps2013labels}
\docType{data}
\title{Results from ski jumps, season 2013/2014}
\description{
Results from ski jumps, season 2013/2014}
\author{
 Data from PAZUR http://estymator.ue.poznan.pl/pazur/
 with small transformations by Przemyslaw Biecek
}
\examples{
head(koty)
}
\keyword{datasets}
