\name{WIG}
\alias{WIG}
\alias{WIG_eng}
\docType{data}
\title{Warszawski Indeks Gieldowy (Warsaw Stock Exchange Index) in 2014}
\description{
WIG, originally an acronym for Warszawski Indeks Gieldowy (Warsaw Stock Exchange Index) is the oldest index of the Warsaw Stock Exchange, introduced on the WSE's first trading session on 16 April 1991.

This data contains daily indexes from year 2014.
}
\author{
 Przemyslaw Biecek
}
\examples{
head(WIG)
head(WIG_eng)
}
\keyword{datasets}
