\name{Diagnoza-Spoleczna}
\alias{diagnoza}
\alias{diagnozaDict}
\docType{data}
\title{A subset of Polish Social Diagnosis data}
\description{
This data is a subset of Diagnoza dataset available on github (https://github.com/pbiecek/Diagnoza).

Source: http://diagnoza.com/
Full dataset: https://github.com/pbiecek/Diagnoza
}
\author{
Przemyslaw Biecek
}
\keyword{package}
\seealso{
}
\examples{
}
