\name{galton}
\alias{galton}
\docType{data}
\title{Galton's height data for parents and children in centimeters}
\description{

This is a data set used by Galton in 1885 to study the relationship between a parent's height and their childrens.
Height is transformed from inches to centimeters.
}
\author{
 Przemyslaw Biecek
}
\examples{
}
\keyword{datasets}
